Use libsigchain on Android
Install page guard signal handler using AddSpecialSignalHandlerFn from
the libsigchain in the Android runtime. This prevents traced
applications from replacing our signal handler when they try to install
thier own signal handlers with sigaction.

This is implicitly enabled when using page_guard on Android and falls
back to sigaction if libsigchain isn't found or the add call fails.

If the signal handler watcher is enabled, the watcher thread will exit
as soon as our handler is successfully added with libsigchain.